Move webhooks endpoints to a place without circular imports
