+++ .golangci.yml
@@ -14,6 +14,7 @@ linters:
     - unparam
     - stylecheck
     - funlen
+    - wsl
 
 linters-settings:
   goimports:
